requirements.txt
Flask==3.0.3
pymongo==4.8.0
python-dotenv==0.19.2
gunicorn==23.0.0